+++ backend/src/main/java/edu/netcracker/jobdealer/service/impl/ResumeServiceImpl.java
@@ -39,7 +39,7 @@ public class ResumeServiceImpl implements ResumeService {
     @Override
     public Resume add(String resumeName, String firstName,
                       String lastName, String about,
-                      String avataUrl, long salary,
+                      String avataUrl, int salary,
                       UUID applicantId, List<String> skillsString)
             throws ApplicantNotFoundException, ResumeAlreadyExistsException {
         Applicant applicant = applicantRepository.findById(applicantId).orElseThrow(ApplicantNotFoundException::new);
